lottie: corrected the comp/layer size data type.
Actually, the these size is expected to be a floating-point value.
This correction addresses a parsing error.

issue: https://github.com/thorvg/thorvg/issues/2153

+++ src/loaders/lottie/tvgLottieParser.h
@@ -55,7 +55,7 @@ private:
 
     void getInperpolatorPoint(Point& pt);
     void getPathSet(LottiePathSet& path);
-    void getLayerSize(uint32_t& val);
+    void getLayerSize(float& val);
     void getValue(TextDocument& doc);
     void getValue(PathSet& path);
     void getValue(Array<Point>& pts);
